+++ testcrate/src/bin/docker_entrypoint_pattern.rs
@@ -21,7 +21,7 @@ use super_orchestrator::{
 use tokio::time::sleep;
 use tracing::info;
 
-const BASE_CONTAINER: &str = "alpine:3.21";
+const BASE_CONTAINER: &str = "alpine:3.23";
 // need this for Alpine
 const TARGET: &str = "x86_64-unknown-linux-musl";